Optionally read the repository-name from cmd-line
Pair programmed with @Vagabond.

With the -r or --repo argument, you can now pass in the
respository to process. This overrides the repo config
option in the bors.cfg file. This is useful if you have a
number of repos that share the same config, save for their
repository name.

The local html and json files have not been updated to reflect
this.