Allow approval comments on the PR if they are followed by a commit SHA
Pair-programmed with @reiddraper

This allows PR comments (which are not editable by the user making the
PR, in contrast to commit comments) to act as approval commends. However
because the SHA is still required, the approval is still tied to a
particular SHA so further pushes to the branch will invalidate it.