Don't manually close PRs after merging
Pair programmed with @Vagabond.

We tested this and github will automatically set the status of the
pull-request to closed/merged if the HEAD of the feature branch is
contained in the target branch. This happens when we fast-forward the
target branch to the merge-sha. And specifically, it looks at the target
branch, not whatever the default branch is set to in github
(master/develop).