Implement a 'no_auto_merge' option
Pair-programmed with @reiddraper

As an option, require bors to see an "@mention merge" comment from a
collaborator before doing the final fast forward of the target branch.

Some Basho engineers are not yet comfortable with PRs being auto-merged.
This allows them to 'get their ducks in a row' before the PR is merged.