Support multiple repositories in bors.html
Pair-programmed with @Vagabond.

bors.html now allows you to choose a specific repository to view, based
on the query string. This is useful if you run bors for multiple repos,
but in the same directory. If there is only one repository, this change
is backward compatible, as it defaults to the singleton repository.
The main change is to wrap the json inside of a map, which is keyed off
of the repository name. The javascript that runs inside the html file
then looks up the queue, based on the query string. If no query-string
is provided (and there is more than one repository), then a list of
available repos is shown.