allow hyphens in usernames
Previously, GitHub usernames containing hyphens were broken: if a
hyphenated username were provided in a `r=some-body` request to bors,
bors would interpret that request as `r=some`.

The reason for this is the regex parsing the username in those comments
was only looking for a word (`\w+`, or equivalently `[a-zA-Z0-9_]+`),
and words end at hyphens.

This changes those regular expressions to look for any of the word
characters and additionally a hyphen character (`[a-zA-Z0-9_-]+`).