Make it possible to ignore comments from certain users
In the case of rust-lang/rust, it'd be rust-highfive. This way,
the bors queue status page will consider PRs from new contributors
"unreviewed". Right now, they're immediately transitioned to the
`DISCUSSING` state due to the initial comment by rust-highfive.
This makes it easy for those PRs to go unnoticed by those contributors
that rely on the queue status to spot new PRs.